Fix #29097 - [Performance] Perf improvement

+++ becpg-core/src/main/java/fr/becpg/repo/repository/impl/AlfrescoRepositoryImpl.java
@@ -1095,7 +1095,9 @@ public class AlfrescoRepositoryImpl<T extends RepositoryEntity> implements Alfre
 	private void removeProperties(NodeRef nodeRef, Set<QName> qnames) {
 		qnames.remove(ContentModel.PROP_NAME);
 		for (QName qname : qnames) {
-			nodeService.removeProperty(nodeRef, qname);
+			if(nodeService.getProperty(nodeRef, qname)!=null) {
+				nodeService.removeProperty(nodeRef, qname);
+			}
 		}
 	}
